# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PR Insight API",
    "description": "우테코 미션 PR 리뷰 코멘트를 자연어로 검색하고, 비슷한 의도의 답변끼리 그룹화하여 요약(2-3문장)과 PR 원문 링크를 함께 제공하는 API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "리뷰 코멘트 검색 (답변 방향성 그룹화)",
        "description": "자연어 query로 conversation document를 검색한 뒤 공식 리뷰어의 개별 답변을 추출하고, 비슷한 방향성의 답변끼리 그룹화한다. 각 그룹(items[])은 화면 인사이트 카드 한 장에 대응하며, 대표 요약과 개별 GitHub comment 아웃링크를 포함한다.",
        "operationId": "search_api_search_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "검색할 자연어 질문",
              "examples": [
                "500에러 예외처리시 응답 구조"
              ],
              "title": "Query"
            },
            "description": "검색할 자연어 질문"
          },
          {
            "name": "track",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "트랙 필터",
              "examples": [
                "BACKEND"
              ],
              "title": "Track"
            },
            "description": "트랙 필터"
          },
          {
            "name": "mission",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "미션 이름 필터",
              "examples": [
                "roomescape"
              ],
              "title": "Mission"
            },
            "description": "미션 이름 필터"
          },
          {
            "name": "repository",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "레포지토리 풀네임 필터",
              "examples": [
                "woowacourse/spring-roomescape-member"
              ],
              "title": "Repository"
            },
            "description": "레포지토리 풀네임 필터"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "그룹화에 사용할 최대 리뷰어 답변 코멘트 수",
              "default": 30,
              "title": "Limit"
            },
            "description": "그룹화에 사용할 최대 리뷰어 답변 코멘트 수"
          },
          {
            "name": "summarize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "true면 GPT로 그룹 제목/요약 생성, false면 대표 snippet 사용",
              "default": true,
              "title": "Summarize"
            },
            "description": "true면 GPT로 그룹 제목/요약 생성, false면 대표 snippet 사용"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/{conversation_id}": {
      "get": {
        "tags": [
          "conversations"
        ],
        "summary": "대화 상세 조회",
        "description": "그룹 카드 클릭 시 호출한다. conversation document에 묶인 원본 리뷰 코멘트들을 시간순(CONTEXT → REPLY)으로 반환하며, 각 코멘트는 GitHub 원문 링크를 포함한다.",
        "operationId": "get_conversation_api_conversations__conversation_id__get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetail"
                }
              }
            }
          },
          "404": {
            "description": "해당 conversation을 찾을 수 없음",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Conversation not found"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/missions": {
      "get": {
        "tags": [
          "filters"
        ],
        "summary": "미션 목록 조회",
        "description": "검색 필터 UI에 노출할 미션 목록을 반환한다. track으로 필터링할 수 있다.",
        "operationId": "list_missions_api_missions_get",
        "parameters": [
          {
            "name": "track",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "특정 트랙의 미션만 조회",
              "examples": [
                "BACKEND"
              ],
              "title": "Track"
            },
            "description": "특정 트랙의 미션만 조회"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MissionOption"
                  },
                  "title": "Response List Missions Api Missions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConversationComment": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role",
            "description": "대화 내 역할: 첫 코멘트는 CONTEXT, 이후는 REPLY.",
            "examples": [
              "CONTEXT"
            ]
          },
          "reviewer": {
            "type": "string",
            "title": "Reviewer",
            "description": "리뷰어 GitHub ID.",
            "examples": [
              "robinjoon"
            ]
          },
          "avatarUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatarurl",
            "examples": [
              "https://avatars.githubusercontent.com/u/12345"
            ]
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "마크다운 본문.",
            "examples": [
              "다른 정보 없이 메세지만 응답하게 설계한 이유가 있을까요?"
            ]
          },
          "githubUrl": {
            "type": "string",
            "title": "Githuburl",
            "description": "이 코멘트의 GitHub 원문 링크.",
            "examples": [
              "https://github.com/woowacourse/spring-roomescape-member/pull/42#discussion_r123"
            ]
          },
          "filePath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filepath",
            "examples": [
              "src/main/java/.../GlobalExceptionHandler.java"
            ]
          },
          "lineNumber": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linenumber",
            "examples": [
              23
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat",
            "examples": [
              "2026-05-21T05:33:09Z"
            ]
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat",
            "examples": [
              "2026-05-21T05:40:00Z"
            ]
          }
        },
        "type": "object",
        "required": [
          "role",
          "reviewer",
          "content",
          "githubUrl",
          "createdAt"
        ],
        "title": "ConversationComment",
        "description": "대화에 포함된 원본 리뷰 코멘트 한 건."
      },
      "ConversationDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "examples": [
              123
            ]
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "examples": [
              "THREAD"
            ]
          },
          "documentText": {
            "type": "string",
            "title": "Documenttext",
            "description": "CONTEXT/REPLY 구분자를 포함한 원본 대화 텍스트.",
            "examples": [
              "[CONTEXT]\nreviewer: robinjoon\n...\n\n---\n\n[REPLY]\nreviewer: yj9107v\n..."
            ]
          },
          "githubUrl": {
            "type": "string",
            "title": "Githuburl",
            "examples": [
              "https://github.com/woowacourse/spring-roomescape-member/pull/42#discussion_r123"
            ]
          },
          "filePath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filepath",
            "examples": [
              "src/main/java/.../GlobalExceptionHandler.java"
            ]
          },
          "lineNumber": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linenumber",
            "examples": [
              23
            ]
          },
          "track": {
            "type": "string",
            "title": "Track",
            "examples": [
              "BACKEND"
            ]
          },
          "mission": {
            "type": "string",
            "title": "Mission",
            "examples": [
              "roomescape"
            ]
          },
          "repository": {
            "type": "string",
            "title": "Repository",
            "examples": [
              "woowacourse/spring-roomescape-member"
            ]
          },
          "prNumber": {
            "type": "integer",
            "title": "Prnumber",
            "examples": [
              42
            ]
          },
          "prTitle": {
            "type": "string",
            "title": "Prtitle",
            "examples": [
              "방탈출 예약 미션 1단계"
            ]
          },
          "comments": {
            "items": {
              "$ref": "#/components/schemas/ConversationComment"
            },
            "type": "array",
            "title": "Comments"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "documentText",
          "githubUrl",
          "track",
          "mission",
          "repository",
          "prNumber",
          "prTitle",
          "comments"
        ],
        "title": "ConversationDetail",
        "description": "대화 상세. 그룹 카드 클릭 시 원본 코멘트들을 시간순으로 반환.",
        "examples": [
          {
            "comments": [
              {
                "avatarUrl": "https://avatars.githubusercontent.com/u/12345",
                "content": "다른 정보 없이 메세지만 응답하게 설계한 이유가 있을까요?",
                "createdAt": "2026-05-21T05:33:09Z",
                "filePath": "src/main/java/.../GlobalExceptionHandler.java",
                "githubUrl": "https://github.com/woowacourse/spring-roomescape-member/pull/42#discussion_r123",
                "lineNumber": 23,
                "reviewer": "robinjoon",
                "role": "CONTEXT"
              },
              {
                "avatarUrl": "https://avatars.githubusercontent.com/u/67890",
                "content": "상태 코드만으로는 구체적 상황 구분이 어려워 errorCode를 함께 내리도록 수정했습니다.",
                "createdAt": "2026-05-21T08:26:36Z",
                "filePath": "src/main/java/.../GlobalExceptionHandler.java",
                "githubUrl": "https://github.com/woowacourse/spring-roomescape-member/pull/42#discussion_r456",
                "lineNumber": 23,
                "reviewer": "yj9107v",
                "role": "REPLY"
              }
            ],
            "documentText": "[CONTEXT]\nreviewer: robinjoon\ncreated_at: 2026-05-21T05:33:09+00:00\n\n다른 정보 없이 메세지만 응답하게 설계한 이유가 있을까요?\n\n---\n\n[REPLY]\nreviewer: yj9107v\ncreated_at: 2026-05-21T08:26:36+00:00\n\n상태 코드만으로는 구체적 상황 구분이 어려워 errorCode를 함께 내리도록 수정했습니다.",
            "filePath": "src/main/java/.../GlobalExceptionHandler.java",
            "githubUrl": "https://github.com/woowacourse/spring-roomescape-member/pull/42#discussion_r123",
            "id": 123,
            "kind": "THREAD",
            "lineNumber": 23,
            "mission": "roomescape",
            "prNumber": 42,
            "prTitle": "방탈출 예약 미션 1단계",
            "repository": "woowacourse/spring-roomescape-member",
            "track": "BACKEND"
          }
        ]
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Conversation not found"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "오류 응답."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MissionOption": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "examples": [
              1
            ]
          },
          "track": {
            "type": "string",
            "title": "Track",
            "examples": [
              "BACKEND"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "roomescape"
            ]
          },
          "repository": {
            "type": "string",
            "title": "Repository",
            "examples": [
              "woowacourse/spring-roomescape-member"
            ]
          },
          "documentCount": {
            "type": "integer",
            "title": "Documentcount",
            "description": "해당 미션의 대화 문서 수.",
            "examples": [
              64
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "track",
          "name",
          "repository",
          "documentCount"
        ],
        "title": "MissionOption",
        "description": "미션 필터 항목."
      },
      "SearchDocument": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "conversation_documents.id. 상세 조회 키.",
            "examples": [
              123
            ]
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "대화 종류: THREAD(질문-답변 스레드), STANDALONE(단독 코멘트), ORPHAN_REPLY(부모 미수집 답변).",
            "examples": [
              "THREAD"
            ]
          },
          "track": {
            "type": "string",
            "title": "Track",
            "examples": [
              "BACKEND"
            ]
          },
          "mission": {
            "type": "string",
            "title": "Mission",
            "examples": [
              "roomescape"
            ]
          },
          "repository": {
            "type": "string",
            "title": "Repository",
            "description": "owner/name 형태.",
            "examples": [
              "woowacourse/spring-roomescape-member"
            ]
          },
          "prNumber": {
            "type": "integer",
            "title": "Prnumber",
            "examples": [
              42
            ]
          },
          "prTitle": {
            "type": "string",
            "title": "Prtitle",
            "examples": [
              "방탈출 예약 미션 1단계"
            ]
          },
          "githubUrl": {
            "type": "string",
            "title": "Githuburl",
            "description": "해당 코멘트의 GitHub 원문 링크([GitHub PR 보기] 버튼).",
            "examples": [
              "https://github.com/woowacourse/spring-roomescape-member/pull/42#discussion_r123"
            ]
          },
          "filePath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filepath",
            "examples": [
              "src/main/java/.../GlobalExceptionHandler.java"
            ]
          },
          "lineNumber": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linenumber",
            "examples": [
              23
            ]
          },
          "reviewers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reviewers",
            "description": "응답 snippet에 포함된 공식 리뷰어 GitHub ID 목록(중복 제거).",
            "examples": [
              [
                "robinjoon"
              ]
            ]
          },
          "snippet": {
            "type": "string",
            "title": "Snippet",
            "description": "공식 리뷰어가 작성한 답변 본문 발췌. 화면에서 검색어 하이라이팅에 사용.",
            "examples": [
              "에러 응답에는 클라이언트가 분기할 수 있는 구조화된 정보가 필요합니다."
            ]
          },
          "score": {
            "type": "number",
            "title": "Score",
            "description": "문서 점수(0~1). vector_score*0.8 + text_score*0.2.",
            "examples": [
              0.87
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "track",
          "mission",
          "repository",
          "prNumber",
          "prTitle",
          "githubUrl",
          "reviewers",
          "snippet",
          "score"
        ],
        "title": "SearchDocument",
        "description": "검색 그룹에 묶인 conversation document 한 건."
      },
      "SearchGroup": {
        "properties": {
          "groupId": {
            "type": "string",
            "title": "Groupid",
            "examples": [
              "cluster-1"
            ]
          },
          "groupTitle": {
            "type": "string",
            "title": "Grouptitle",
            "description": "그룹 대표 제목. summarize=true면 GPT 생성, 아니면 대표 문서 파일명/키워드 기반.",
            "examples": [
              "에러 응답 구조"
            ]
          },
          "representativeAnswer": {
            "type": "string",
            "title": "Representativeanswer",
            "description": "그룹을 대표하는 2-3문장 요약/답변. summarize=true면 GPT 요약, 아니면 대표 snippet.",
            "examples": [
              "에러 메시지만 내려주면 프론트엔드에서 상황별 분기 처리가 어렵습니다. errorCode나 HTTP 상태 코드 기반으로 응답을 구조화하라는 리뷰가 다수입니다. 클라이언트가 에러 유형을 명확히 구분하도록 하기 위함입니다."
            ]
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "그룹에 묶인 리뷰어 답변 수.",
            "examples": [
              5
            ]
          },
          "score": {
            "type": "number",
            "title": "Score",
            "description": "그룹 대표 점수(대표 답변 기준).",
            "examples": [
              0.87
            ]
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/SearchDocument"
            },
            "type": "array",
            "title": "Documents",
            "description": "그룹에 묶인 문서 목록(상위 5개)."
          },
          "reviewerSections": {
            "items": {
              "$ref": "#/components/schemas/SearchReviewerSection"
            },
            "type": "array",
            "title": "Reviewersections",
            "description": "답변 방향성 그룹에 포함된 리뷰어별 개별 코멘트 목록."
          }
        },
        "type": "object",
        "required": [
          "groupId",
          "groupTitle",
          "representativeAnswer",
          "count",
          "score",
          "documents",
          "reviewerSections"
        ],
        "title": "SearchGroup",
        "description": "비슷한 의도의 답변끼리 묶인 인사이트 카드 한 장."
      },
      "SearchResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SearchGroup"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "SearchResponse",
        "description": "검색 응답. items는 화면 인사이트 카드 목록에 대응.",
        "examples": [
          {
            "items": [
              {
                "count": 5,
                "documents": [
                  {
                    "filePath": "src/main/java/.../GlobalExceptionHandler.java",
                    "githubUrl": "https://github.com/woowacourse/spring-roomescape-member/pull/42#discussion_r123",
                    "id": 123,
                    "kind": "THREAD",
                    "lineNumber": 23,
                    "mission": "roomescape",
                    "prNumber": 42,
                    "prTitle": "방탈출 예약 미션 1단계",
                    "repository": "woowacourse/spring-roomescape-member",
                    "reviewers": [
                      "robinjoon"
                    ],
                    "score": 0.87,
                    "snippet": "에러 응답에는 클라이언트가 분기할 수 있는 구조화된 정보가 필요합니다.",
                    "track": "BACKEND"
                  }
                ],
                "groupId": "cluster-1",
                "groupTitle": "에러 응답 구조",
                "representativeAnswer": "에러 메시지만 내려주면 프론트엔드에서 상황별 분기 처리가 어렵습니다. errorCode나 HTTP 상태 코드 기반으로 응답을 구조화하라는 리뷰가 다수입니다. 클라이언트가 에러 유형을 명확히 구분하도록 하기 위함입니다.",
                "reviewerSections": [
                  {
                    "commentCount": 1,
                    "comments": [
                      {
                        "commentGithubId": 123456789,
                        "content": "에러 응답에는 클라이언트가 분기할 수 있는 구조화된 정보가 필요합니다.",
                        "conversationId": 123,
                        "createdAt": "2026-05-21T08:26:36Z",
                        "documentGithubUrl": "https://github.com/woowacourse/spring-roomescape-member/pull/42#discussion_r123",
                        "filePath": "src/main/java/.../GlobalExceptionHandler.java",
                        "githubUrl": "https://github.com/woowacourse/spring-roomescape-member/pull/42#discussion_r456",
                        "lineNumber": 23,
                        "prNumber": 42,
                        "prTitle": "방탈출 예약 미션 1단계",
                        "repository": "woowacourse/spring-roomescape-member",
                        "score": 0.87,
                        "snippet": "에러 응답에는 클라이언트가 분기할 수 있는 구조화된 정보가 필요합니다."
                      }
                    ],
                    "nickname": "로빈",
                    "reviewer": "robinjoon"
                  }
                ],
                "score": 0.87
              }
            ]
          }
        ]
      },
...
      "SearchReviewerSection": {
        "properties": {
          "reviewer": {
            "type": "string",
            "title": "Reviewer",
            "description": "리뷰어 GitHub ID.",
            "examples": [
              "robinjoon"
            ]
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname",
            "description": "리뷰어 닉네임.",
            "examples": [
              "로빈"
            ]
          },
          "commentCount": {
            "type": "integer",
            "title": "Commentcount",
            "description": "해당 리뷰어의 코멘트 수.",
            "examples": [
              2
            ]
          },
          "comments": {
            "items": {
              "$ref": "#/components/schemas/SearchReviewerComment"
            },
            "type": "array",
            "title": "Comments"
          }
        },
        "type": "object",
        "required": [
          "reviewer",
          "commentCount",
          "comments"
        ],
        "title": "SearchReviewerSection",
        "description": "답변 방향성 그룹 안에서 리뷰어별로 묶은 코멘트 목록."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}